# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Deep Research AI Agent API",
    "description": "API for conducting comprehensive research using AI agents",
    "version": "1.0.0"
  },
  "paths": {
    "/api/generate-questions": {
      "post": {
        "operationId": "generateQuestions",
        "summary": "Generate clarifying questions for a research topic",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "topic": {
                    "type": "string",
                    "description": "The research topic to generate questions for"
                  }
                },
                "required": [
                  "topic"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully generated clarifying questions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/foundry/research": {
      "post": {
        "operationId": "runDeepResearch",
        "summary": "Run deep research on a topic (synchronous)",
        "description": "Runs research synchronously and waits for completion. Returns the full report and activities.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "topic": {
                    "type": "string",
                    "description": "The research topic or question to investigate"
                  },
                  "clarifications": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "question": {
                          "type": "string"
                        },
                        "answer": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "question",
                        "answer"
                      ]
                    }
                  }
                },
                "required": [
                  "topic"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Research completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "report": {
                      "type": "string",
                      "nullable": true
                    },
                    "activities": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "timestamp": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/foundry/research/async": {
      "post": {
        "operationId": "runDeepResearchAsync",
        "summary": "Start deep research asynchronously (returns session ID immediately)",
        "description": "Kicks off research in the background and returns immediately with a session ID. Track progress by querying SessionEvent objects filtered by sessionId.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "topic": {
                    "type": "string",
                    "description": "The research topic or question to investigate"
                  },
                  "clarifications": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "question": {
                          "type": "string"
                        },
                        "answer": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "question",
                        "answer"
                      ]
                    }
                  }
                },
                "required": [
                  "topic"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Research started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "sessionId": {
                      "type": "string",
                      "description": "Unique ID to track research progress via SessionEvent objects"
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "success",
                    "sessionId"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}